+++ tests/BlazorServerFunctions.Generator.UnitTests/_snapshots/GrpcServerGeneratorTests/GrpcServerGeneratorTests.Generate_BasicGrpcInterface_ProducesServiceClass#IUserServiceGrpcService.g.verified.cs
@@ -4,7 +4,7 @@
 
 using ProtoBuf;
 using ProtoBuf.Grpc;
-using ProtoBuf.Grpc.Configuration;
+using System.ServiceModel;
 using System.Threading.Tasks;
 
 namespace MyApp.Services;
@@ -18,14 +18,12 @@ public sealed class UserServiceGetUserAsyncGrpcRequest
     public string Name { get; set; } = default!;
 }
 
-[ServiceContract]
-public sealed class UserServiceGrpcService
+public sealed class UserServiceGrpcService : IUserServiceGrpcContract
 {
     private readonly IUserService _service;
 
     public UserServiceGrpcService(IUserService service) => _service = service;
 
-    [OperationContract]
     public Task<string> GetUserAsync(UserServiceGetUserAsyncGrpcRequest request, CallContext context = default)
         => _service.GetUserAsync(request.Id, request.Name);