+++ tests/BlazorServerFunctions.Generator.UnitTests/_snapshots/GrpcServerGeneratorTests/GrpcServerGeneratorTests.Generate_MixedRestAndGrpc_BothGenerated#IGrpcServiceGrpcService.g.verified.cs
@@ -4,7 +4,7 @@
 
 using ProtoBuf;
 using ProtoBuf.Grpc;
-using ProtoBuf.Grpc.Configuration;
+using System.ServiceModel;
 using System.Threading.Tasks;
 
 namespace MyApp.Services;
@@ -16,14 +16,12 @@ public sealed class GrpcServiceCallAsyncGrpcRequest
     public int Id { get; set; }
 }
 
-[ServiceContract]
-public sealed class GrpcServiceGrpcService
+public sealed class GrpcServiceGrpcService : IGrpcServiceGrpcContract
 {
     private readonly IGrpcService _service;
 
     public GrpcServiceGrpcService(IGrpcService service) => _service = service;
 
-    [OperationContract]
     public Task<string> CallAsync(GrpcServiceCallAsyncGrpcRequest request, CallContext context = default)
         => _service.CallAsync(request.Id);